# openapi.yaml
openapi: 3.0.0
info:
  title: Trello Clone API
  description: Backend for a simple Trello clone application
  version: 1.0.0
servers:
  - url: /
    description: Server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  schemas:
    Board:
      type: object
      required:
        - name
      properties:
        id:
          type: string
        name:
          type: string
        userId:
          type: string
        createdAt:
          type: string
          format: date-time
        isArchived:
          type: boolean
          default: false
        background:
          type: string
          description: URL or color code for board background
        isTemplate:
          type: boolean
          default: false
        isFavorite:
          type: boolean
          default: false
        members:
          type: array
          items:
            type: object
            properties:
              userId:
                type: string
              role:
                type: string
                enum: [owner, admin, member]
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
        password:
          type: string
          format: password
    LoginResponse:
      type: object
      required:
        - token
      properties:
        token:
          type: string
          description: JWT token for authentication
    User:
      type: object
      required:
        - username
      properties:
        id:
          type: string
        username:
          type: string
        createdAt:
          type: string
          format: date-time
    RegisterRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 30
        password:
          type: string
          minLength: 6
          format: password
    LogoutResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
...
security:
  - bearerAuth: []
paths:
  /sessions:
    post:
      tags:
        - Authentication
      security: []  # No authentication required for login
      summary: Create a session to get JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              username: "newUser"
              password: "securePassword"
      responses:
        200:
          description: Successful session creation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        401:
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags:
        - Authentication
      security:
        - bearerAuth: []
      summary: End the current session and revoke the token
      responses:
        200:
          description: Successfully logged out
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutResponse'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users:
    get:
      tags:
        - Users
      security:
        - bearerAuth: []
      summary: Retrieve all users
      responses:
        200:
          description: A list of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - Users
      security: []  # No authentication required for creating a user
      summary: Create a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            example:
              username: "newUser"
              password: "securePassword"
      responses:
        201:
          description: User created successfully
          headers:
            Location:
              schema:
                type: string
              description: URL of the created user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        400:
          description: Invalid username or password format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        409:
          description: Username already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags:
        - Users
      security:
        - bearerAuth: []
      summary: Delete current user's account
      responses:
        204:
          description: User deleted successfully
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /boards:
    get:
      tags:
        - Boards
      security:
        - bearerAuth: []  # Require authentication for GET
      summary: Get all boards
      responses:
        200:
          description: A list of boards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Board'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - Boards
      security:
        - bearerAuth: []
      summary: Create a new board
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 100
            example:
              name: "New Board"
      responses:
        201:
          description: Board created
          headers:
            Location:
              schema:
                type: string
              description: URL of the created board
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Board'
        400:
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /boards/{boardId}:
    get:
      tags:
        - Boards
      security:
        - bearerAuth: []
      summary: Get a specific board
      parameters:
        - in: path
          name: boardId
          required: true
          schema:
            type: string
      responses:
        200:
          description: Board retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Board'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        403:
          description: Forbidden - Not authorized to view this board
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        404:
          description: Board not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "Board not found."
    put:
      tags:
        - Boards
      security:
        - bearerAuth: []
      summary: Update a board
      parameters:
        - in: path
          name: boardId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 100
                background:
                  type: string
                isTemplate:
                  type: boolean
                isFavorite:
                  type: boolean
                isArchived:
                  type: boolean
            example:
              name: "Updated Board Name"
              background: "#FFFFFF"
              isTemplate: false
              isFavorite: false
              isArchived: false
      responses:
        200:
          description: Board updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Board'
        400:
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        403:
          description: Forbidden - Not authorized to update this board
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        404:
          description: Board not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags:
        - Boards
      security:
        - bearerAuth: []
      summary: Delete a board
      parameters:
        - in: path
          name: boardId
          required: true
          schema:
            type: string
      responses:
        204:
          description: Board deleted successfully
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        403:
          description: Forbidden - Not authorized to delete this board
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        404:
          description: Board not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /cards/{cardId}/checklist:
    post:
      tags:
        - Cards
      security:
        - bearerAuth: []
      summary: Add checklist to card
      parameters:
        - in: path
          name: cardId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
              properties:
                title:
                  type: string
            example:
              title: "New Checklist"
      responses:
        201:
          description: Checklist added successfully
          headers:
            Location:
              schema:
                type: string
              description: URL of the created checklist
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Checklist added successfully"
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...